backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.48
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.11
alembic==1.12.0
pytest==7.4.0
httpx==0.24.1
python-dotenv==1.0.0
bcrypt==3.2.2
email-validator==1.3.1
python-multipart==0.0.6
